Change bigint to integer in Postgres schema grammar.
Bigint has too big a range to be returned as an int in PHP. It is returned as a string, which can be confusing and lead to unexpected results. In addition, the range of integer is suitable for most cases: -2147483648 to +2147483647.

+++ src/Illuminate/Database/Schema/Grammars/PostgresGrammar.php
@@ -219,7 +219,7 @@ class PostgresGrammar extends Grammar {
 	 */
 	protected function typeInteger(Fluent $column)
 	{
-		return $column->autoIncrement ? 'serial' : 'bigint';
+		return $column->autoIncrement ? 'serial' : 'integer';
 	}
 
 	/**
